requirements.txt
openai
requests
schedule
PyYAML
google-api-python-client
google-auth
